12257: Выделил методы получения статистики в отдельный класс CgroupStats
